Fix button style and wallet balance mix
